refactor: extract ts-render-analysis, migrate video_analysis — 27/27 (#91)
Финальная группа cutover. video_analysis была заблокирована анализным поддеревом
`video_compiler::core::ffmpeg` (scene/motion/quality/silence/keyframes/audio/...), которое
осталось в монолите И циклически ссылалось на саму группу (возвращало её типы-результаты).

Разрыв цикла + вынос:
- НОВЫЙ крейт `ts-render-analysis` (deps: ts-render, ts-render-services, ts-analysis):
  10 анализных модулей из core/ffmpeg/ + перенесённые СЮДА типы-результаты
  (VideoMetadata/Scene/QualityAnalysisResult/KeyFrame/SilenceSegment/Motion*/Audio* —
  бывш. video_analysis/types.rs). Теперь и движок, и команда импортируют типы из крейта —
  цикл разорван.
  Пути переписаны: super::FFmpegCommand -> re-export из ts-render-services;
  crate::video_compiler::{error,core::error} -> ts_render::...; crate::analysis::types ->
  ts_analysis::analysis::types; crate::video_compiler::commands::video_analysis::<типы> -> crate::types.
- Монолит core/ffmpeg/mod.rs -> ре-экспорт-шим `pub use ts_render_analysis::*` (потребители
  analysis::services::{scene_detector,unified_audio_analyzer}, state::commands::media — без правок).
- Группа `video_analysis` -> ts-render-tauri: core::ffmpeg::* -> ts_render_analysis::*,
  state -> ts_render_services, error/ffmpeg_executor -> ts_render; types.rs -> шим на ts_render_analysis::types.
  app_builder: 8 команд квалифицированы `video_analysis::`.

ИТОГО: 27 из 27 групп video_compiler/commands вынесены из монолита в крейты. 🎉

Проверки: cargo test -p ts-render-analysis -> 22 passed; cargo test -p ts-render-tauri
--test-threads=2 -> 538 passed (8 ignored=интеграционные); cargo check --lib (src-tauri) ->
0 ошибок, 3 baseline-варнинга (без новых).
