+++ crates/ts-render-analysis/src/audio_analysis.rs
@@ -1,10 +1,10 @@
 //! Модуль для анализа аудио с помощью FFmpeg
 
 use super::FFmpegCommand;
-use crate::video_compiler::commands::video_analysis::{
+use crate::types::{
   AudioAnalysisResult, AudioQuality, DynamicsData, FrequencyData, VolumeData,
 };
-use crate::video_compiler::error::{Result, VideoCompilerError};
+use ts_render::video_compiler::error::{Result, VideoCompilerError};
 use regex::Regex;
 use std::path::Path;
