+++ crates/ts-render-analysis/src/types.rs
