+++ crates/ts-render-tauri/src/video_analysis/types.rs
@@ -1,0 +1,3 @@
+//! Типы результатов анализа — вынесены в крейт `ts-render-analysis` (#91).
+//! Ре-экспорт-шим, чтобы `super::types::*` в commands/business_logic/tests резолвился.
+pub use ts_render_analysis::types::*;
