+++ src-tauri/src/app_builder.rs
@@ -328,14 +328,14 @@ pub fn build_app<R: Runtime>() -> Builder<R> {
     crate::video_compiler::commands::monitoring::reset_all_metrics,
     crate::video_compiler::commands::monitoring::get_registry_service_metrics,
     // Video analysis commands
-    crate::video_compiler::commands::ffmpeg_get_metadata,
-    crate::video_compiler::commands::ffmpeg_detect_scenes,
-    crate::video_compiler::commands::ffmpeg_analyze_quality,
-    crate::video_compiler::commands::ffmpeg_detect_silence,
-    crate::video_compiler::commands::ffmpeg_analyze_motion,
-    crate::video_compiler::commands::ffmpeg_extract_keyframes,
-    crate::video_compiler::commands::ffmpeg_analyze_audio,
-    crate::video_compiler::commands::ffmpeg_quick_analysis,
+    crate::video_compiler::commands::video_analysis::ffmpeg_get_metadata,
+    crate::video_compiler::commands::video_analysis::ffmpeg_detect_scenes,
+    crate::video_compiler::commands::video_analysis::ffmpeg_analyze_quality,
+    crate::video_compiler::commands::video_analysis::ffmpeg_detect_silence,
+    crate::video_compiler::commands::video_analysis::ffmpeg_analyze_motion,
+    crate::video_compiler::commands::video_analysis::ffmpeg_extract_keyframes,
+    crate::video_compiler::commands::video_analysis::ffmpeg_analyze_audio,
+    crate::video_compiler::commands::video_analysis::ffmpeg_quick_analysis,
     // Whisper commands
     crate::video_compiler::commands::whisper_commands::whisper_transcribe_openai,
     crate::video_compiler::commands::whisper_commands::whisper_translate_openai,
